+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Array output parameters now return `system.Array` instead of `list`.
 - Fill arrays now take a buffer as the argument and fill that buffer instead
   of taking a size as the argument and returning a new array.
+- Use `typing.ClassVar` annotation for static properties.
 
 ### Fixed
 - Fixed checking wrong number of input parameters on methods with array parameters.